+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `CallableTree::Node::Internal#broadcast` (experimental)
+
 ## [0.1.0] - 2021-05-19
 
 - Initial release